% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_to.R
\docType{methods}
\name{codon_to}
\alias{codon_to}
\title{Maximize or minimize the usage of certain codon.}
\usage{
codon_to(object, max.codon = NA, min.codon = NA, ...)

\S4method{codon_to}{regioned_dna}(object, max.codon, min.codon)
}
\arguments{
\item{object}{A regioned_dna object.}

\item{max.codon}{A string of a codon.}

\item{min.codon}{A string of a codon.}

\item{...}{}
}
\value{
A regioned_dna object.
}
\description{
Input string of a codon to either the "max.codon = " or "min.codon = "
parameter to maximize or minimize the usage of certain codon in the sequence.
}
\examples{
get_cu(codon_to(rgd.seq, max.codon = "AAC")) - get_cu(rgd.seq)
get_cu(codon_to(rgd.seq, min.codon = "AAC")) - get_cu(rgd.seq)

}
\seealso{
\code{\link{input_seq}}, \code{\link{dinu_to}},
  \code{\link{codon_random}}, \code{\link{codon_mimic}}
}
