% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage.R
\docType{methods}
\name{codon_usage}
\alias{codon_usage}
\title{codon_usage}
\usage{
codon_usage(object, region = NA, ...)

\S4method{codon_usage}{character}(object, region)

\S4method{codon_usage}{DNAStringSet}(object, region)
}
\arguments{
\item{\code{object}}{filepath or DNAstringSet}

\item{\code{region}}{a data-frame specifying paticular regions that is
allowed to be mutated in the sequences. Both \code{0 / 1} or
\code{TRUE / FALSE} encoding is OK. Please refer to Examples below for
reference.}
}
\value{
A codon_usage_data-class object
}
\description{
Constructing \code{codon_usage_data} from DNAStringSet.
}
\examples{
# Creating a codon_usage class directly from system file
filepath <- system.file("extdata", "example.fasta", package = "SynMut")
cu <- codon_usage(filepath)

#Optionally input with region dataframe
filepath <- system.file("extdata", "example.fasta", package = "SynMut")
filepath.csv <- system.file("extdata", "target_regions.csv", package = "SynMut")
region <- read.csv(filepath.csv)
cu <- codon_usage(filepath, region)

# Creating from exsisting DNAStringSet object
seq <- Biostrings::DNAStringSet("ATCGATCGA")
cu <- codon_usage(seq)
}
\seealso{
\code{\link{get_cu}}, \code{\link{get_region}}
}
