% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage.R
\docType{methods}
\name{codon_usage}
\alias{codon_usage}
\alias{codon_usage,character-method}
\alias{codon_usage,DNAStringSet-method}
\title{codon_usage}
\usage{
codon_usage(object, ...)

\S4method{codon_usage}{character}(object)

\S4method{codon_usage}{DNAStringSet}(object)
}
\arguments{
\item{...}{filepath or DNAstringSet}
}
\value{
A codon_usage_data-class object
}
\description{
Constructing \code{codon_usage_data} from DNAStringSet.
}
\examples{
# Creating a codon_usage class directly from system file
filepath <- system.file("extdata", "example.fasta", package = "SynMut")
cu <- codon_usage(filepath)

# Creating from exsisting DNAStringSet object
seq <- Biostrings::DNAStringSet("ATCGATCGA")
cu <- codon_usage(seq)
}
\seealso{
\code{\link{get_cu}}
}
