% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinu_to.R
\docType{methods}
\name{dinu_to}
\alias{dinu_to}
\title{Maximize or minimize the usage of certain dinucleotide.}
\usage{
dinu_to(object, max.dinu = NA, min.dinu = NA, ...)

\S4method{dinu_to}{regioned_dna}(object, max.dinu, min.dinu)
}
\arguments{
\item{object}{A regioned_dna object.}

\item{max.dinu}{A string of a dinucleotide.}

\item{min.dinu}{A string of a dinucleotide.}
}
\value{
regioned_dna
}
\description{
Input string of a dinucleotide to either the "max.dinu = " or "min.codon = "
parameter to maximize or minimize the usage of certain codon in the sequence.
Using a greedy algorithm with priority given to dinucleotide12 or
dinucleotide23.
}
\examples{
get_du(dinu_to(rgd.seq, max = "cg")) - get_du(rgd.seq)
get_du(dinu_to(rgd.seq, min = "aa")) - get_du(rgd.seq)
}
\seealso{
\code{\link{input_seq}}, \code{\link{codon_to}},
  \code{\link{codon_random}}, \code{\link{codon_mimic}},
  \code{\link{codon_mimic_dinu_to}}
}
