% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioned_dna_Class.R
\docType{methods}
\name{get_cu}
\alias{get_cu}
\title{Get codon usage matrix}
\usage{
get_cu(object, ...)

\S4method{get_cu}{regioned_dna}(object)

\S4method{get_cu}{DNAStringSet}(object)
}
\arguments{
\item{object}{regioned_dna / DNAStringSet}
}
\value{
matrix
}
\description{
Access the codon usage matrix
}
\examples{
get_cu(rgd.seq)
}
\seealso{
\code{\link{input_seq}}, \code{\link{get_region}},
  \code{\link{get_nu}}, \code{\link{get_du}}, \code{\link{get_freq}},
  \code{\link{get_rscu}}
}
