% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioned_dna_Class.R
\docType{methods}
\name{get_dna}
\alias{get_dna}
\alias{get_dna,regioned_dna-method}
\title{Get the DNAStringSet data}
\usage{
get_dna(object, ...)

\S4method{get_dna}{regioned_dna}(object)
}
\arguments{
\item{object}{A regioned_dna object.}

\item{...}{...}
}
\value{
DNAStringSet
}
\description{
Access the DNA sequence data in DNAStringSet.
}
\examples{
filepath <- system.file("extdata", "example.fasta", package = "SynMut")
rgd.seq <- input_seq(filepath)
get_dna(rgd.seq)
}
