% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioned_dna_Class.R
\docType{methods}
\name{get_du}
\alias{get_du}
\title{Get dinucleotide usage matrix}
\usage{
get_du(object, ...)

\S4method{get_du}{regioned_dna}(object)

\S4method{get_du}{DNAStringSet}(object)
}
\arguments{
\item{object}{regioned_dna / DNAStringSet}
}
\value{
matrix
}
\description{
Access the dinucleotide usage matrix
}
\examples{
get_du(rgd.seq)
}
\seealso{
\code{\link{input_seq}}, \code{\link{get_region}},
  \code{\link{get_cu}}, \code{\link{get_freq}}, \code{\link{get_rscu}}
}
