% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioned_dna_Class.R
\docType{methods}
\name{get_freq}
\alias{get_freq}
\title{Get codon usage frequency of synonymous codons}
\usage{
get_freq(object, ...)

\S4method{get_freq}{regioned_dna}(object)

\S4method{get_freq}{DNAStringSet}(object)

\S4method{get_freq}{matrix}(object)

\S4method{get_freq}{vector}(object)
}
\arguments{
\item{object}{regioned_dna / DNAStringSet / codon usage matrix (vector)}
}
\value{
matrix
}
\description{
Access the synonymous codon usage frequency
}
\examples{
get_freq(rgd.seq)
}
\seealso{
\code{\link{input_seq}}, \code{\link{get_region}},
  \code{\link{get_cu}}, \code{\link{get_du}}, \code{\link{get_rscu}}
}
