% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioned_dna_Class.R
\docType{methods}
\name{get_nu}
\alias{get_nu}
\title{Get nucleotide usage matrix}
\usage{
get_nu(object, ...)

\S4method{get_nu}{regioned_dna}(object)

\S4method{get_nu}{DNAStringSet}(object)
}
\arguments{
\item{object}{regioned_dna / DNAStringSet}
}
\value{
matrix
}
\description{
Access the nucleotide usage matrix
}
\examples{
get_nu(rgd.seq)
}
\seealso{
\code{\link{input_seq}}, \code{\link{get_region}},
  \code{\link{get_cu}}, \code{\link{get_du}}, \code{\link{get_rscu}}
}
