% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage_data_Class.R
\docType{methods}
\name{get_region}
\alias{get_region}
\title{get_region: get the variable region}
\usage{
get_region(object, ...)

\S4method{get_region}{codon_usage_data}(object)
}
\arguments{
\item{\code{cu}}{codon_usage_data}
}
\value{
list
}
\description{
access the variable regions
}
\examples{
codon.matrix <- get_region(cu)
}
\seealso{
\code{\link{codon_usage}}, \code{\link{get_cu}}
}
