% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioned_dna_Class.R
\docType{methods}
\name{get_region}
\alias{get_region}
\alias{get_region,regioned_dna-method}
\title{Get the variable region}
\usage{
get_region(object, ...)

\S4method{get_region}{regioned_dna}(object)
}
\arguments{
\item{object}{regioned_dna}

\item{...}{...}
}
\value{
list
}
\description{
Access the variable regions
}
\examples{
filepath <- system.file("extdata", "example.fasta", package = "SynMut")
rgd.seq <- input_seq(filepath)
get_region(rgd.seq)
}
\seealso{
\code{\link{input_seq}}, \code{\link{get_cu}}
}
