% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioned_dna_Class.R
\docType{methods}
\name{get_rscu}
\alias{get_rscu}
\title{Get Relative Synonymous Codon Usage (rscu) of synonymous codons}
\usage{
get_rscu(object, ...)

\S4method{get_rscu}{regioned_dna}(object)

\S4method{get_rscu}{DNAStringSet}(object)
}
\arguments{
\item{object}{regioned_dna / DNAStringSet / codon usage matrix (vector)}
}
\value{
matrix
}
\description{
Access the Relative Synonymous Codon Usage rscu
}
\examples{
get_rscu(rgd.seq)
}
\seealso{
\code{\link{input_seq}}, \code{\link{get_region}},
  \code{\link{get_cu}}, \code{\link{get_du}}, \code{\link{get_freq}}
}
