% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioned_dna_Class.R
\docType{class}
\name{regioned_dna-class}
\alias{regioned_dna-class}
\title{An S4 class to record DNA sequences and variable regions for mutations}
\description{
Recording codon DNA sequences and region.
}
\section{Slots}{

\describe{
\item{\code{\code{dnaseq}}}{a DNAStingSet object recording the sequence(s)}

\item{\code{\code{region}}}{a list specifying paticular regions in the sequences
allowed to be mutated}
}}

\seealso{
\code{\link{input_seq}}, \code{\link{get_cu}},
  \code{\link{get_region}}
}
\author{
Haogao Gu
}
