% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_mimic.R
\docType{methods}
\name{codon_mimic}
\alias{codon_mimic}
\alias{codon_mimic,regioned_dna,vector-method}
\alias{codon_mimic,regioned_dna,DNAStringSet-method}
\title{Mimic a target codon usage bias}
\usage{
codon_mimic(object, alt, ...)

\S4method{codon_mimic}{regioned_dna,vector}(object, alt)

\S4method{codon_mimic}{regioned_dna,DNAStringSet}(object, alt)
}
\arguments{
\item{object}{regioned_dna object}

\item{alt}{target codon usage vector or DNAStringSet object representing
target codon usage}

\item{...}{...}
}
\value{
regioned_dna
}
\description{
Mutating the current DNA sequences in the regioned_dna object to mimic a
target codon usage pattern.
}
\details{

}
\examples{
filepath <- system.file("extdata", "example.fasta", package = "SynMut")
rgd.seq <- input_seq(filepath)
target <- get_cu(rgd.seq)[2,]
new <- codon_mimic(rgd.seq, alt = target)
get_cu(new) - get_cu(rgd.seq)

target <- Biostrings::DNAStringSet("TTGAAAA-CTC-N--AAG")
new <- codon_mimic(rgd.seq, alt = target)
get_cu(new) - get_cu(rgd.seq)
get_freq(new) - get_freq(rgd.seq)
get_rscu(new) - get_rscu(rgd.seq)
}
\seealso{
\code{\link{input_seq}}, \code{\link{codon_to}},
  \code{\link{codon_random}}, \code{\link{dinu_to}}
}
