% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_random.R
\docType{methods}
\name{codon_random}
\alias{codon_random}
\alias{codon_random,regioned_dna-method}
\title{Generate random synonymous mutations}
\usage{
codon_random(object, n = NA, keep = FALSE, ...)

\S4method{codon_random}{regioned_dna}(object, n, keep)
}
\arguments{
\item{object}{A regioned_dna object.}

\item{n}{Optional n parameter specifying what proportion of the codons
to be mutate. Default value: \code{NA}.}

\item{keep}{Logical parameter controling whether keeping the codon
usage bias of the original sequence. Default value: \code{FALSE}.}

\item{...}{...}
}
\value{
A regioned_dna object containing the mutants
}
\description{
Generating radom synonymous mutations (in user-defined region), with
optionally keeping/not keeping the original codon usage bias.
}
\examples{
filepath <- system.file("extdata", "example.fasta", package = "SynMut")
rgd.seq <- input_seq(filepath)
set.seed(2019)
get_cu(codon_random(rgd.seq, n = 0.5))
get_cu(codon_random(rgd.seq))
}
\seealso{
\code{\link{input_seq}}, \code{\link{dinu_to}},
  \code{\link{codon_to}}, \code{\link{codon_mimic}}
}
